refactor(src): add build system
